requirements.txt
annotated-types==0.6.0
certifi==2023.7.22
charset-normalizer==3.3.2
idna==3.4
pydantic==2.5.0
pydantic_core==2.14.1
python-dateutil==2.8.2
PyYAML==6.0.1
requests==2.31.0
rss-parser==1.2.1
six==1.16.0
typing_extensions==4.8.0
urllib3==2.1.0
xmltodict==0.13.0
